Completed test create/edit logic
