+++ src/Pages/Home/modules/NewTest/index.ts
@@ -1,1 +1,1 @@
-export * from './CreateTest';
+export * from './TestDrawer';
