+++ src/Pages/Home/utils/getQuestionAmount.ts
