--- src/Pages/HomeTEST/modules/AddButton/index.ts
@@ -1,1 +1,0 @@
-export * from './AddButton';
