--- src/Pages/HomeTEST/modules/NewTest/index.ts
@@ -1,1 +1,0 @@
-export * from './TestDrawer';
