--- src/Pages/HomeTEST/modules/NewTest/modules/index.ts
@@ -1,1 +1,0 @@
-export * from './SubDrawer';
