--- src/Pages/HomeTEST/modules/TestList/index.ts
@@ -1,1 +1,0 @@
-export * from './TestList';
