--- src/Pages/HomeTEST/modules/TestListItem/index.ts
@@ -1,1 +1,0 @@
-export * from './TestListItem';
