--- src/Pages/HomeTEST/modules/index.ts
@@ -1,3 +1,0 @@
-export * from './TestListItem';
-export * from './TestList';
-export * from './AddButton';
