+++ src/hooks/index.ts
@@ -3,4 +3,3 @@ export * from './useBoolean';
 export * from './useForm';
 export * from './useQuestions';
 export * from './useTests';
-export * from './useImage';
