+++ task.md
@@ -1,7 +1,7 @@
 # Quiz Tasks
 
-- Добавление/изменение картинки
 - Стилизация Session
+  <!-- - Добавление/изменение картинки -->
   <!-- - Добавить изменение вопроса (только текст) -->
   <!-- - Исправить баг с прокруткой в drawer (список вопросов) -->
   <!-- - Необходим useEffect для обновления вопроса в subDrawer -->
@@ -9,6 +9,7 @@
 
 ## Оптимизация
 
+- Контейнер для ui - max-width
 - Делать запрос на получение только одного теста при измении/создании, а не целого списка всех тестов
 - Уменьшение кол-ва типов (переиспользование)
 - Уменьшение кол-ва перерендеров
